+++ ongabot/handler/helpcommandhandler.py
@@ -4,6 +4,7 @@ from telegram import Update
 from telegram.ext import CallbackContext, CommandHandler
 
 from utils import helper
+from utils.dm import is_private_chat
 from utils.log import log
 
 
@@ -17,4 +18,4 @@ class HelpCommandHandler(CommandHandler):
 @log
 async def callback(update: Update, _: CallbackContext) -> None:
     """Print the help text for a /start or /help command"""
-    await update.message.reply_text(helper.create_help_text())
+    await update.message.reply_text(helper.create_help_text(private=is_private_chat(update)))